# api-docs/_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rent System API",
    "description": "Rent System API Description",
    "contact": {
      "name": "Rent"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/v1/dummy": {
      "parameters": [
        {
          "$ref": "#/components/parameters/acceptLanguage"
        }
      ],
      "get": {
        "summary": "Dummy get",
        "tags": [
          "Dummy"
        ],
        "operationId": "get-dummy",
        "description": "Dummy get",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dummyGetRes"
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common400"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common500"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/field"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "security": [
          {
            "tokenAuth": []
          }
        ]
      },
      "post": {
        "summary": "Dummy register",
        "tags": [
          "Dummy"
        ],
        "operationId": "post-dummy",
        "description": "Dummy register",
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid parameter supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common400"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dummyPos409tRes"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common500"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dummyPostRes"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Dummy edit",
        "tags": [
          "Dummy"
        ],
        "operationId": "patch-dummy",
        "description": "Dummy edit",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid parameter supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common400"
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common404"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dummyPos409tRes"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common500"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dummyPostRes"
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Dummy delete",
        "tags": [
          "Dummy"
        ],
        "operationId": "delete-dummy",
        "description": "Dummy delete",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid parameter supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common400"
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common404"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common500"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/emailReq"
          }
        ],
        "security": [
          {
            "tokenAuth": []
          }
        ]
      }
    },
...
    "/v1/auth/refresh-token": {
      "parameters": [
        {
          "$ref": "#/components/parameters/acceptLanguage"
        }
      ],
      "post": {
        "summary": "refresh-token",
        "tags": [
          "Auth"
        ],
        "operationId": "refresh-token",
        "description": "Refreshing refresh and access token",
        "responses": {
          "200": {
            "description": "Tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/refreshGenerated200"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Invalid Token / Token has expired / Wrong token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/unauthorized401"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common500"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/refreshTokenReq"
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "tokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authorization token that a client provides when making API calls. (e.g. Authorization: Bearer <Token>)"
      }
    },
    "schemas": {
...
      "refreshGenerated200": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "object",
            "properties": {
              "access_token": {
                "type": "string",
                "example": "token......"
              },
              "refresh_token": {
                "type": "string",
                "example": "token......"
              }
            },
            "required": [
              "refresh_token",
              "access_token"
            ]
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
      "common400": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Not valid"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "unauthorized401": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Otp not valid."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "common404": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Data not found"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "common500": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Internal server error"
          }
        },
        "required": [
          "message"
        ]
      },
      "dummyGetRes": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "object",
            "properties": {
              "totalCount": {
                "type": "integer",
                "example": 2
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 4
                    },
                    "name": {
                      "type": "string",
                      "example": "test2"
                    },
                    "email": {
                      "type": "string",
                      "example": "test2@test.com"
                    },
                    "description": {
                      "type": "string",
                      "example": "test2"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-12-09T17:12:15.000Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-12-09T17:12:15.000Z"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "email"
                  ]
                }
              }
            },
            "required": [
              "totalCount",
              "data"
            ]
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "dummyPostRes": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "dummy"
          },
          "email": {
            "type": "string",
            "example": "dummy@dummy.com"
          },
          "description": {
            "type": "string",
            "example": "Dummy description"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "dummyPos409tRes": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Data already exist"
          },
          "errorData": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "dummy@dummy.com"
              }
            }
          }
        },
        "required": [
          "message",
          "errorData"
        ]
      },
...
      "refreshTokenReq": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string",
            "example": "*******"
          }
        },
        "required": [
          "device_id"
        ]
      }
    },
    "parameters": {
      "acceptLanguage": {
        "schema": {
          "type": "string",
          "enum": [
            "hi",
            "en"
          ]
        },
        "in": "header",
        "name": "accept-language",
        "description": "To support API localization",
        "required": false
      },
...
      "emailReq": {
        "schema": {
          "type": "string"
        },
        "in": "query",
        "name": "email",
        "example": "dummy@dummy.com",
        "description": "email",
        "required": true
      },
...
      "field": {
        "schema": {
          "type": "string"
        },
        "in": "query",
        "name": "email",
        "example": "test@test.com | like",
        "description": "you can parse any field like <fieldValue> | like OR <fieldValue> | equals"
      },
...
      "limit": {
        "schema": {
          "type": "string"
        },
        "in": "query",
        "name": "limit",
        "example": "10",
        "description": "how many record you want in one page"
      },
      "offset": {
        "schema": {
          "type": "string"
        },
        "in": "query",
        "name": "offset",
        "example": "1",
        "description": "page number"
      },
      "order": {
        "schema": {
          "type": "string"
        },
        "in": "query",
        "name": "order",
        "example": "email | asc",
        "description": "you can parse ordering like <fieldName> | asc OR <fieldName> | desc"
      },
...
    }
  }
}